Create Environment.swift
